Move task graph related code to its own module task_graph
